Add support for git-worktree style repositories
This change adds support for git-worktree style repositories. The
majority of information kept in GitInfo is inferred through invocation
of `git` itself. The program has proper support for git-worktree style
repositories already and so nothing changes there. However, special
handling is needed for the _giFiles member which tracks git files. To
compile this list of files we check whether <root>/.git is a directory
or not and in case it is not (as is the case in a git-worktree style
repository), we merely add the .git file to this list.