Bugfixes: Let `getGitFilesRegular` work with branching and rewording
Fixed several bugs in `getGitFilesRegular`.

1. It now stores `./git/HEAD` as dependency even if you don't have your
HEAD detached.

2. It now strips the trailing newlines from the content of `.git/HEAD`
to get the ref file of the current branch.

With this commit the function does both. Prior to this commit,
commit/ref-level changes that doesn't affect the actual working
directory was often ignored. With the absence of the Item 1,
`git branch -b` was dismissed and not considered as a change by GHC.
Without the Item 2, `git commit --amend --only`
directory was ignored by GHC.